Module docstrings. Fixed tests. Moved calc_layer_props back to atmosphere module. Moved colours to style module. Add header to all files
