+++ src/compose/ocean.jl
@@ -1,3 +1,4 @@
+# This file is part of AGNI. License is GPL-3.0: https://www.gnu.org/licenses
 # Contains module handling surface oceans
 
 """
