+++ src/interface/setpt.jl
@@ -1,12 +1,14 @@
-# Contains functions for setting-up the temperature profile analytically
+# This file is part of AGNI. License is GPL-3.0: https://www.gnu.org/licenses
 
+"""
+**Contains functions for setting-up the temperature profile analytically**
+"""
 module setpt
 
     import ..phys
     import ..species
     import ..atmosphere
     import ..chemistry
-    import ..layers
     import ..guillot
 
     using NCDatasets
@@ -162,7 +164,7 @@ module setpt
             end
 
             # Calculate height (etc) but don't error if it fails, yet
-            if !layers.calc_layer_props!(atmos)
+            if !atmosphere.calc_layer_props!(atmos)
                 @warn "Initial T(p) structure may be unbound"
             end
 
@@ -483,7 +485,7 @@ module setpt
         atmos.tmpl[end] = atmos.tmp_surf
 
         # Set mmw
-        layers.calc_profile_mmw!(atmos)
+        atmosphere.calc_profile_mmw!(atmos)
 
         # Lapse rate dT/dp
         grad::Float64 = 0.0
@@ -496,8 +498,8 @@ module setpt
             if i < atmos.nlev_c
                 atmos.tmp[i] = atmos.tmp[i+1]
             end
-            layers.calc_single_cpkc!(atmos, i)
-            layers.calc_single_density!(atmos, i)
+            atmosphere.calc_single_cpkc!(atmos, i)
+            atmosphere.calc_single_density!(atmos, i)
 
             # Evaluate lapse rate dT/dp
             grad = 1 / (atmos.layer_ρ[i] * atmos.layer_cp[i])
