+++ test/test_integration.jl
@@ -97,7 +97,7 @@ TEST_DIR        = joinpath(ROOT_DIR,"test/")
         atmosphere.allocate!(atmos,"$RES_DIR/stellar_spectra/sun.txt")
         setpt.dry_adiabat!(atmos)
         setpt.saturation!(atmos, "H2O")
-        layers.calc_layer_props!(atmos)
+        atmosphere.calc_layer_props!(atmos)
         energy.radtrans!(atmos, true)
         energy.radtrans!(atmos, false)
 
@@ -305,6 +305,7 @@ TEST_DIR        = joinpath(ROOT_DIR,"test/")
                                 real_gas=false
                         )
         atmosphere.allocate!(atmos,"$RES_DIR/stellar_spectra/sun.txt")
+        atmosphere.calc_layer_props!(atmos)
         energy.radtrans!(atmos, true)
         energy.radtrans!(atmos, false)
 
@@ -350,7 +351,7 @@ TEST_DIR        = joinpath(ROOT_DIR,"test/")
                         )
         atmosphere.allocate!(atmos,"$RES_DIR/stellar_spectra/sun.txt")
         setpt.isothermal!(atmos, 300.0)
-        layers.calc_layer_props!(atmos)
+        atmosphere.calc_layer_props!(atmos)
         atmos.flux_tot[:] .= 0.0
         energy.radtrans!(atmos, true)
         energy.radtrans!(atmos, false)
